Создан компонент карточек
